4.homework/calculator/src/main/java/com/solncev/servlets/CalculatorServlet.java
package com.solncev.servlets;

import com.solncev.services.CalculatorService;

import javax.servlet.ServletException;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

/**
 * Created by Марат on 03.12.2016.
 */
public class CalculatorServlet extends HttpServlet {

    @Override
    public void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
        String request = req.getRequestURI();
        resp.getWriter().print(new CalculatorService().getAnswer(request));
    }

    @Override
    public void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
        resp.setStatus(405);
    }
}
